library: implement uptime change without breaking API
This commit represents a partial reversal of the patch
submitted by D. Cantrell adding a pretty print option
to the uptime program.  It also, hopefully, will serve
as an example of how to minimize the impact of changes
to our brand new libprocps.

As originally implemented, the patch necessarily
required changes to the library as well as the uptime
program.  However, it also necessitated modifications
to the top and w programs and possiblly other unknown
programs since the API had been changed.

After this commit, the library API and ABI will remain
unchanged.

Reference(s):
commit 3e7f78d296e610922daa7e144f52a77392705563

Signed-off-by: Jim Warner <james.warner@comcast.net>